+++ standalone/src/data_conversion_layer/start_request_serialization.cpp
@@ -80,14 +80,15 @@ RawData data_conversion_layer::start_request::serialize(const data_conversion_la
    */
 
   const uint8_t device_enabled{ 0b00001000 };
-  const uint8_t intensity_enabled{ 0b00001000 };
+  const uint8_t intensity_enabled{ static_cast<uint8_t>(
+      msg.master_device_settings_.intensitiesEnabled() ? 0b00001000 : 0b00000000) };
   const uint8_t point_in_safety_enabled{ 0 };
   const uint8_t active_zone_set_enabled{ 0 };
   const uint8_t io_pin_enabled{ 0 };
   const uint8_t scan_counter_enabled{ 0b00001000 };
   const uint8_t speed_encoder_enabled{ 0 }; /**< 0000000bin disabled, 00001111bin enabled.*/
   const uint8_t diagnostics_enabled{ static_cast<uint8_t>(
-      msg.master_device_settings_.isDiagnosticsEnabled() ? 0b00001000 : 0b00000000) };
+      msg.master_device_settings_.diagnosticsEnabled() ? 0b00001000 : 0b00000000) };
 
   raw_processing::write(os, device_enabled);
   raw_processing::write(os, intensity_enabled);